RetryQueue If Condtion
Included an "if" condition if there is any called queued for retries. If a retry is active, don't call the other method, and focus on the call that needs to be made. This seem to remove the bug where 1 dialogue is waiting, and another can not proceed to the next dialogue until the other one is done delaying. This will need to undergo further test to confirm this.