Cast_Enters_From macro
Introduced a new macro to have the cast enter from the left or right side of the screen.

You must call this macro before calling the "cue_cast" macro. To see the effect of the cast member entering from the left or right side of the screen to the center,  use the "set_cast_motion" macro, and set it to "Interpolation". 

This is almost at a "beta" state, and can be used in production.