KeycodeDefinitions Implemented
You can assign a VirtualKey and provide it an InputEvent. You can use the InputEvent to see if any of the VirtualKeys falls into that event.

For example, the PROCEED InputEvent is called when you progress to the next dialogue. You can assign multiple VirtualKeys just as Return or LeftMouse, and XVNML2U will act upon that.