Updated "XVNMLModule"
You can now allow for cache usage and generation of the XVNMLModule. This is an option for those not entirely worried about game performance. This is if the XVNML file itself is relatively small. However, if you're XVNML get's big, enabling this will ensure that it doesn't create a whole new XVNMLObj to use, but instead use the cache.json file that was generated for it.